Pagination adapter improvement.
