--- packages/components/src/EntityList/pagination/index.ts
@@ -1,1 +1,0 @@
-export { default as EntityPagination } from './EntityPagination';
