--- packages/components/src/EntityList/pagination/style.scss
@@ -1,29 +1,0 @@
-/** -------------------------------------------------------------------------
- * Event Espresso EntityPagination Stylesheet
- * © 2020 Event Espresso
- * ------------------------------------------------------------------------- */
-
-@import '~@eventespresso/styles/src/mixins/breakpoints.scss';
-
-.ee-entity-pagination {
-	display: flex;
-	flex-flow: row nowrap;
-	justify-content: flex-end;
-	padding: var(--ee-padding-default);
-
-	@include max782px {
-		padding: var(--ee-padding-small);
-	}
-
-	@include max600px {
-		padding: var(--ee-padding-smaller);
-	}
-
-	@include max480px {
-		padding: var(--ee-padding-tiny);
-	}
-
-	@include max360px {
-		padding: var(--ee-padding-micro);
-	}
-}
