--- packages/components/src/EntityList/pagination/types.ts
@@ -1,12 +1,0 @@
-import type { Entity } from '@eventespresso/data';
-
-export type onChangeFn = (page: number) => void;
-
-export type onShowSizeChangeFn = (current: number, size: number) => void;
-
-export interface PaginationProps<E extends Entity> {
-	onChange: onChangeFn;
-	onShowSizeChange: onShowSizeChangeFn;
-	paginatedEntities: E[];
-	total: number;
-}
