+++ packages/components/src/Pagination/style.scss
